renaming: keygen->utils
The name was actually badly chosen at first, we aren't doing any key
generation but only providing some utility functions to help with
creating SEAL parameters and context to build key generators.
