+++ tests/test_enc_dec.py
